pkg/client/logging: Disolve stat_unix.go in to stat_{darwin,linux}.go
