+++ pkg/client/logging/stat_linux.go
@@ -1,8 +1,32 @@
 package logging
 
 import (
+	"fmt"
+	"os"
+	"syscall"
 	"time"
 )
+
+type fileInfo struct {
+	*syscall.Stat_t
+}
+
+func GetSysInfo(_ string, info os.FileInfo) (SysInfo, error) {
+	return fileInfo{info.Sys().(*syscall.Stat_t)}, nil
+}
+
+func (u fileInfo) SetOwnerAndGroup(name string) error {
+	return os.Chown(name, int(u.Uid), int(u.Gid))
+}
+
+func (u fileInfo) HaveSameOwnerAndGroup(other SysInfo) bool {
+	ou := other.(fileInfo)
+	return u.Uid == ou.Uid && u.Gid == ou.Gid
+}
+
+func (u fileInfo) String() string {
+	return fmt.Sprintf("CTIME %v, UID %d, GID %d", u.BirthTime(), u.Uid, u.Gid)
+}
 
 func (u fileInfo) BirthTime() time.Time {
 	return time.Unix(u.Ctim.Sec, u.Ctim.Nsec)
